backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pillow
